cleanup

+++ hyu/src/wl/wp_presentation_feedback.rs
@@ -31,8 +31,8 @@ impl WpPresentationFeedback {
 			object_id: *self.object_id,
 			op: 1,
 			args: (
-				((time.as_secs() as u64 >> 32) & 0xFFFFFFFF) as u32,
-				((time.as_secs() as u64) & 0xFFFFFFFF) as u32,
+				((time.as_secs() >> 32) & 0xFFFFFFFF) as u32,
+				(time.as_secs() & 0xFFFFFFFF) as u32,
 				(((time.as_nanos() % 1_000_000_000) as u64) & 0xFFFFFFFF) as u32,
 				till_next_refresh,
 				((sequence >> 32) & 0xFFFFFFFF) as u32,
